Checkpoint for new Changes / Sleep
Fixing Controllers and Routing. Going to get Upload working next. It didn't specify in requirements if I needed to make a UI to consume endpoints, but I want this to come out well and get as good with MVC Core as Hibernate and work towards become more full stack in my next role. Should be done with requirements tomorrow but the entire project I intend to complete could take a while.